+++ frontend/components/authAction/useLogin.ts
@@ -33,23 +33,10 @@ import { loggedInUserState } from '@/recoil/atom';
 import { useRouter } from 'next/router';
 import { api } from '@/util/api';
 import { setCookie } from '@/util/cookie';
+import { LoginData } from '@/types/user';
 
-interface IObj {
-  isLoading: boolean;
-  data: undefined | any;
-  auth: undefined | any;
-  error: undefined | any;
-}
-
-const BASE_URL = 'http://13.209.14.135:8080/';
-
-export default function usePostApi(
-  endpoint: string
-): [(data?: any) => void, IObj] {
+export default function usePostApi(endpoint: string) {
   const [isLoading, setIsLoading] = useState(false);
-  const [data, setData] = useState<undefined | any>(undefined);
-  const [auth, setAuth] = useState<undefined | any>(undefined);
-  const [error, setError] = useState<undefined | any>(undefined);
 
   //유저 데이터 저장할 atom
   const setUser = useSetRecoilState(loggedInUserState);
@@ -57,24 +44,29 @@ export default function usePostApi(
   //조회에 성공하면 이동하기 위해
   const router = useRouter();
 
-  function mutation(data?: any) {
+  function Login(data: LoginData) {
     setIsLoading(true);
-    // axios
-    //   .post(BASE_URL + endpoint, data)
     api
       .post(endpoint, data)
       .then((res) => {
         setCookie('accessToken', res.headers['authorization'], 40); //로그인 했을 때, 쿠키 설정
         setUser(res.data); //로그인한 유저의 데이터를 atom 관리
-        setData(res);
-        setAuth(res.headers['authorization']);
       })
       .then(() => {
-        router.push('/'); //값을 모두 세팅하고 홈으로 이동
+        router.push('/');
       })
-      .catch(setError)
       .finally(() => setIsLoading(false));
   }
 
-  return [mutation, { isLoading, data, auth, error }];
+  function SignUp(data: LoginData) {
+    setIsLoading(true);
+    api
+      .post(endpoint, data)
+      .then(() => {
+        router.push('/users/login'); //회원가입에 성공하면 로그인으로 이동
+      })
+      .finally(() => setIsLoading(false));
+  }
+
+  return [Login, SignUp];
 }
