+++ frontend/types/user.ts
@@ -26,3 +26,10 @@ export type UserState = {
   totalStar: number;
   yearOfDev: number;
 };
+
+//로그인 또는 회원가입 할때 데이터 타입
+export type LoginData = {
+  name?: string;
+  email: string;
+  password: string | number;
+};
